third_party/WebKit/LayoutTests/platform/win/external/wpt/IndexedDB/idbcursor_update_objectstore9-expected.txt

This is a testharness.js-based test.
PASS IDBCursor.update() - object store - throw InvalidStateError when the cursor is being iterated 
Harness: the test ran to completion.
